Strip the `sz` image URL query parameters
Fixes #166 once again since Google started injecting
query parameters with `?z=50` to return a default size of 50
with all profile image URLs even when an explicit size is requested.

This means that even when we receive a URL like:

https://lh6.googleusercontent.com/-lDP57qXndCU/AAAAAAAAAAI/AAAAAAAEeVQ/M3c432SPq4I/s160-c/photo.jpg?sz=50

It will be converted to:
https://lh6.googleusercontent.com/-lDP57qXndCU/AAAAAAAAAAI/AAAAAAAEeVQ/M3c432SPq4I/s160-c/photo.jpg

But URLs with other query parameters will not be affected as requested
by @zquestz.