Added tweak to Omniauth callback example
Long time, man.  We should catch up.

I was just pulling this into a project and found that the large oauth payload can overflow cookie sizes if you use cookies as your sesion store.  In general, it is probably not needed to store the extra data either, so added code to remove that and comment it out.