UID should always be sub
According to -> https://developers.google.com/identity/protocols/OpenIDConnect#server-flow
(scroll down a bit)

"An identifier for the user, unique among all Google accounts and never reused. A Google account can have multiple emails at different points in time, but the sub value is never changed. Use sub within your application as the unique-identifier key for the user."

It is also always passed back. Or was that || because of backwards compatibility?