Do not verify iat claim for JWT (#392)
In https://github.com/jwt/ruby-jwt/commit/a1ede264b3ae0e3178686bcc4d2302983e566838,
ruby-jwt removed the ability to allow a leeway when verifying the iat claim.
That means that even tiny skew in server clocks can cause a JWT::InvalidIatError to be raised

According to the JWT specification at https://tools.ietf.org/html/rfc7519#section-4.1.6
the iat claim is not intended to be verified. If a JWT is only valid after
a certain time, the nbf (not before) claim will be used.